Delete dubious double-download
I can't imagine the purpose of this code that starts, and then cancels,
a second download in parallel with the first. After recent changes to
the internals of dl-tar, the cancellation was no longer effective, and
the two download-and-extract processes ended up interfering with each
other sometimes.

Removing this code seems strictly beneficial; maybe one day I'll regret
this but right now I can't see a reason to keep it.